re-structuring functions
